+++ src/beryll/core/EventHandler.h
@@ -36,7 +36,7 @@ namespace Beryll
         // True if finger was down at this game loop iteration.
         // Will false in next game loop iteration (reset in loadEvents()).
         bool downEvent;
-        long ID;
+        int ID;
     };
 
     class EventHandler
